+++ packages/server/src/lib.rs
@@ -5,9 +5,11 @@
 mod config;
 mod document;
 mod error;
+mod ext;
 mod freshness;
 #[cfg(not(target_arch = "wasm32"))]
 mod fs_cache;
+pub use ext::*;
 pub mod launch;
 mod memory_cache;
 pub use error::*;
